fix: added new helper file for sonar cloud issue
